subindo testes userService

+++ src/main/java/br/com/catalisa/ZuPlaceApi/service/UserService.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 
@@ -19,35 +20,50 @@ public class UserService {
     @Autowired
     private ModelMapper mapper;
 
-    public List<UserModel> findAll(){
+    public List<UserModel> findAll() {
         return repository.findAll();
     }
-    public UserModel findById(Long id){
-        return  repository.findById(id).orElseThrow(() -> new ResourseNotFoundException("usuário não encontrado"));
+
+    public UserModel findById(Long id) {
+        Optional<UserModel> user = repository.findById(id);
+        return user.orElseThrow(() -> new ResourseNotFoundException("usuário não encontrado"));
     }
 
-    public UserResponseDto create( UserRequestDto userRequestDto){
+    public UserModel create(UserResponseDto userResponseDto) {
 
-        try{
+        try {
 
-            UserModel userModel = mapper.map(userRequestDto, UserModel.class);
-            userModel = repository.save(userModel);
-UserResponseDto responseDto = mapper.map(userModel, UserResponseDto.class);
-return  responseDto;
-        } catch (Exception e){
-            throw  new ResourseNotFoundException("usuário não encontrado");
+            UserModel userModel = mapper.map(userResponseDto, UserModel.class);
+
+            return repository.save(userModel);
+        } catch (Exception e) {
+            throw new ResourseNotFoundException("usuário não cadastrado");
         }
 
     }
 
-    public UserModel update (UserResponseDto userResponseDto){
-        return repository.save(mapper.map(userResponseDto, UserModel.class));
+    public UserModel update(Long id, UserResponseDto userResponseDto) {
+        UserModel updatedUser  = findById(id);
+        if (updatedUser == null){
+            throw  new ResourseNotFoundException("Usuário não encontrado");
+        }
+        updatedUser.setName(userResponseDto.getName());
+updatedUser.setEmail(userResponseDto.getEmail());
+updatedUser.setPassword(userResponseDto.getPassword());
+updatedUser.setPersonType(userResponseDto.getPersonType());
+updatedUser.setPhone(userResponseDto.getPhone());
+updatedUser.setDocumentType(userResponseDto.getDocumentType());
+return  repository.save(updatedUser);
     }
 
-    public void delete(Long id){
+    public void delete(Long id) {
         UserModel validateId = findById(id);
-        repository.deleteById(validateId.getId());
+        if (validateId != null) {
+            repository.deleteById(validateId.getId());
+        } else {
+            throw new ResourseNotFoundException("usuário não encontrado");
+        }
     }
-}
 
+}
